add userAuth to BookingController

+++ src/main/java/com/sctp/module3project2/controller/BookingController.java
@@ -2,6 +2,8 @@ package com.sctp.module3project2.controller;
 
 import java.util.ArrayList;
 
+import javax.naming.AuthenticationException;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -14,7 +16,9 @@ import com.sctp.module3project2.entity.Booking;
 import com.sctp.module3project2.services.BookingService;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.PutMapping;
+import com.sctp.module3project2.services.UseridService;
 
 
 // Joel
@@ -23,39 +27,61 @@ import org.springframework.web.bind.annotation.PutMapping;
 @RequestMapping("/api/booking")
 public class BookingController {
     private BookingService bookingService;
+    private UseridService useridService;
     
-    public BookingController(BookingService bookingService){
+    public BookingController(BookingService bookingService, UseridService useridService){
         this.bookingService = bookingService;
+        this.useridService = useridService;
     }
 
     @GetMapping("")
-    public ResponseEntity<ArrayList<Booking>> getAllBookings(){
+    public ResponseEntity<ArrayList<Booking>> getAllBookings(@RequestHeader(value="password") String password, @RequestHeader(value="user") String user) throws AuthenticationException{
+        String passwordSystem = useridService.findPassWordByUserID(user);
+        if (!(passwordSystem.equals(password))){
+            throw new AuthenticationException("Username/Password authentication not accepted");
+        }
         ArrayList<Booking> allBookings = bookingService.getAllBookings();
         return ResponseEntity.ok(allBookings);
 
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Booking> getBooking(@PathVariable Long id){
+    public ResponseEntity<Booking> getBooking(@PathVariable Long id, @RequestHeader(value="password") String password, @RequestHeader(value="user") String user) throws AuthenticationException{
+        String passwordSystem = useridService.findPassWordByUserID(user);
+        if (!(passwordSystem.equals(password))){
+            throw new AuthenticationException("Username/Password authentication not accepted");
+        }
         Booking foundBooking = bookingService.getBooking(id);
         return ResponseEntity.ok(foundBooking);
     }
 
     @PostMapping("")
-    public ResponseEntity<Booking> saveBooking(@RequestBody Booking booking){
+    public ResponseEntity<Booking> saveBooking(@RequestBody Booking booking, @RequestHeader(value="password") String password, @RequestHeader(value="user") String user) throws AuthenticationException{
+        String passwordSystem = useridService.findPassWordByUserID(user);
+        if (!(passwordSystem.equals(password))){
+            throw new AuthenticationException("Username/Password authentication not accepted");
+        }
         Booking newBooking = bookingService.saveBooking(booking);
         return ResponseEntity.ok(newBooking);
     }
 
     @PutMapping(value="/{id}")
-    public ResponseEntity<Booking> updateBooking(@PathVariable Long id, @RequestBody Booking booking){
+    public ResponseEntity<Booking> updateBooking(@PathVariable Long id, @RequestBody Booking booking, @RequestHeader(value="password") String password, @RequestHeader(value="user") String user) throws AuthenticationException{
+        String passwordSystem = useridService.findPassWordByUserID(user);
+        if (!(passwordSystem.equals(password))){
+            throw new AuthenticationException("Username/Password authentication not accepted");
+        }
         Booking updatedBooking = bookingService.updateBooking(id, booking);
         return ResponseEntity.ok(updatedBooking);
     }
 
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Booking> deleteBooking(@PathVariable Long id){
+    public ResponseEntity<Booking> deleteBooking(@PathVariable Long id , @RequestHeader(value="password") String password, @RequestHeader(value="user") String user) throws AuthenticationException{
+        String passwordSystem = useridService.findPassWordByUserID(user);
+        if (!(passwordSystem.equals(password))){
+            throw new AuthenticationException("Username/Password authentication not accepted");
+        }
         bookingService.deleteBooking(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
